feat(6695): extract cli scripted inline tests
