+++ crates/kamn-e2e-harness/src/drivers/cli_scripted.rs
@@ -2112,1367 +2112,5 @@ fn parse_text_output_field<'a>(output: &'a str, key: &str) -> Option<&'a str> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::env;
-    use super::{
-        live_execution_enabled_from_env, parse_s15_budget_env_u128, parse_text_output_field,
-        run_cli_command_capture_stdout, run_live_s01_cli_health_probe,
-        run_live_s02_cli_direct_message_probe, run_live_s03_cli_group_channel_probe,
-        run_live_s04_cli_task_lifecycle_probe, run_live_s05_cli_escrow_settlement_probe,
-        run_live_s06_cli_proof_verification_probe, run_live_s07_cli_replay_protection_probe,
-        run_live_s08_cli_crash_recovery_probe, run_live_s09_cli_transport_failover_probe,
-        run_live_s10_cli_topology_coherence_probe, run_live_s11_cli_signer_rotation_probe,
-        run_live_s12_cli_retention_deletion_probe, run_live_s13_cli_bridge_forwarding_probe,
-        run_live_s14_cli_batch_merkle_probe, run_live_s15_cli_performance_smoke_probe,
-        validate_live_s05_release_escrow_response, validate_s07_replay_reason_marker,
-        validate_s08_message_receipt_fields, validate_s08_query_message_response,
-        validate_s12_content_field_coherence, validate_s12_content_id_match,
-        validate_s13_bridge_field_coherence, validate_s13_bridge_id_match,
-        validate_s14_cli_verify_proof_response, validate_s15_latency_budget_samples,
-        CliScriptedDriver, CLI_BINARY_ENV, CLI_SCRIPTED_LIVE_ENV,
-    };
-    use std::ffi::OsString;
-    use std::fs;
-    use std::os::unix::fs::PermissionsExt;
-    use std::path::PathBuf;
-    use std::sync::PoisonError;
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    fn with_env_vars<F>(updates: &[(&str, Option<&str>)], test: F)
-    where
-        F: FnOnce(),
-    {
-        let _guard = crate::drivers::test_env_lock()
-            .lock()
-            .unwrap_or_else(PoisonError::into_inner);
-        let previous = updates
-            .iter()
-            .map(|(key, _)| ((*key).to_owned(), env::var_os(key)))
-            .collect::<Vec<(String, Option<OsString>)>>();
-
-        for (key, value) in updates {
-            match value {
-                Some(value) => {
-                    // SAFETY: tests serialize env mutation with a process-wide mutex.
-                    unsafe { env::set_var(key, value) }
-                }
-                None => {
-                    // SAFETY: tests serialize env mutation with a process-wide mutex.
-                    unsafe { env::remove_var(key) }
-                }
-            }
-        }
-
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(test));
-        for (key, value) in previous {
-            match value {
-                Some(value) => {
-                    // SAFETY: tests serialize env mutation with a process-wide mutex.
-                    unsafe { env::set_var(key, value) }
-                }
-                None => {
-                    // SAFETY: tests serialize env mutation with a process-wide mutex.
-                    unsafe { env::remove_var(key) }
-                }
-            }
-        }
-        if let Err(payload) = result {
-            std::panic::resume_unwind(payload);
-        }
-    }
-
-    fn unique_temp_script_path(stem: &str) -> PathBuf {
-        let nonce = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("system time should be after unix epoch")
-            .as_nanos();
-        std::env::temp_dir().join(format!("{stem}-{}-{nonce}.py", std::process::id()))
-    }
-
-    fn write_executable_python_script(script_path: &PathBuf, source: &str) {
-        fs::write(script_path, source).expect("script fixture should be written");
-        let mut permissions = fs::metadata(script_path)
-            .expect("script metadata should be readable")
-            .permissions();
-        permissions.set_mode(0o755);
-        fs::set_permissions(script_path, permissions).expect("script fixture should be executable");
-    }
-
-    #[test]
-    fn unit_parse_bool_flag_accepts_true_like_values() {
-        for value in ["1", "true", "TRUE", "yes", "on"] {
-            assert!(
-                crate::drivers::shared_helpers::parse_bool_flag(value),
-                "expected truthy for {value}"
-            );
-        }
-    }
-
-    #[test]
-    fn unit_parse_bool_flag_rejects_false_like_values() {
-        for value in ["0", "false", "off", "no", ""] {
-            assert!(
-                !crate::drivers::shared_helpers::parse_bool_flag(value),
-                "expected falsey for {value}"
-            );
-        }
-    }
-
-    #[test]
-    fn unit_live_execution_enabled_from_env_honors_true_and_false_markers() {
-        with_env_vars(&[(CLI_SCRIPTED_LIVE_ENV, Some("1"))], || {
-            assert!(live_execution_enabled_from_env());
-        });
-        with_env_vars(&[(CLI_SCRIPTED_LIVE_ENV, Some("0"))], || {
-            assert!(!live_execution_enabled_from_env());
-        });
-    }
-
-    #[test]
-    fn unit_run_live_s01_cli_health_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error =
-                    run_live_s01_cli_health_probe().expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s01_cli_health_probe_sets_deterministic_identity_opt_in_env() {
-        let script_path = unique_temp_script_path("kamn-e2e-cli-s01-deterministic-opt-in");
-        let script_source = r#"#!/usr/bin/env python3
-import os
-import sys
-
-command = sys.argv[1] if len(sys.argv) > 1 else ""
-if command != "health":
-    sys.stderr.write("unexpected command")
-    sys.exit(2)
-
-if os.environ.get("KAMN_AGENT_LIB_ALLOW_DETERMINISTIC_IDENTITY") != "1":
-    sys.stderr.write("deterministic identity opt-in missing")
-    sys.exit(3)
-
-sys.stdout.write("status=ok")
-"#;
-        write_executable_python_script(&script_path, script_source);
-
-        with_env_vars(
-            &[
-                (
-                    CLI_BINARY_ENV,
-                    Some(
-                        script_path
-                            .to_str()
-                            .expect("script path should be valid utf-8"),
-                    ),
-                ),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-                ("KAMN_AGENT_LIB_ALLOW_DETERMINISTIC_IDENTITY", None),
-            ],
-            || {
-                run_live_s01_cli_health_probe()
-                    .expect("probe should opt in deterministic identity for cli child process");
-            },
-        );
-
-        fs::remove_file(&script_path).expect("script fixture should be removable");
-    }
-
-    #[test]
-    fn unit_run_live_s02_cli_direct_message_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s02_cli_direct_message_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s03_cli_group_channel_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error =
-                    run_live_s03_cli_group_channel_probe().expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s03_cli_group_channel_probe_rejects_query_message_id_mismatch() {
-        let script_path = unique_temp_script_path("kamn-e2e-cli-s03-query-mismatch");
-        let script_source = r#"#!/usr/bin/env python3
-import sys
-
-command = sys.argv[1] if len(sys.argv) > 1 else ""
-if command == "create-channel":
-    sys.stdout.write("channel_id=channel-1 status=created")
-elif command == "send-message":
-    sys.stdout.write("message_id=message-1 status=sent")
-elif command == "query-message":
-    sys.stdout.write("message_id=message-2 status=sent")
-elif command == "list-messages":
-    sys.stdout.write("channel_id=channel-1 messages=[message-1]")
-else:
-    sys.stderr.write("unsupported command")
-    sys.exit(2)
-"#;
-        write_executable_python_script(&script_path, script_source);
-
-        with_env_vars(
-            &[
-                (
-                    CLI_BINARY_ENV,
-                    Some(
-                        script_path
-                            .to_str()
-                            .expect("script path should be valid utf-8"),
-                    ),
-                ),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s03_cli_group_channel_probe()
-                    .expect_err("mismatched query message_id should fail");
-                assert!(
-                    error.contains("mismatched message_id"),
-                    "error should mention message_id mismatch: {error}",
-                );
-            },
-        );
-
-        fs::remove_file(&script_path).expect("script fixture should be removable");
-    }
-
-    #[test]
-    fn unit_run_live_s03_cli_group_channel_probe_rejects_list_channel_id_mismatch() {
-        let script_path = unique_temp_script_path("kamn-e2e-cli-s03-list-mismatch");
-        let script_source = r#"#!/usr/bin/env python3
-import sys
-
-command = sys.argv[1] if len(sys.argv) > 1 else ""
-if command == "create-channel":
-    sys.stdout.write("channel_id=channel-1 status=created")
-elif command == "send-message":
-    sys.stdout.write("message_id=message-1 status=sent")
-elif command == "query-message":
-    sys.stdout.write("message_id=message-1 status=sent")
-elif command == "list-messages":
-    sys.stdout.write("channel_id=channel-2 messages=[message-1]")
-else:
-    sys.stderr.write("unsupported command")
-    sys.exit(2)
-"#;
-        write_executable_python_script(&script_path, script_source);
-
-        with_env_vars(
-            &[
-                (
-                    CLI_BINARY_ENV,
-                    Some(
-                        script_path
-                            .to_str()
-                            .expect("script path should be valid utf-8"),
-                    ),
-                ),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s03_cli_group_channel_probe()
-                    .expect_err("mismatched listed channel_id should fail");
-                assert!(
-                    error.contains("mismatched channel_id"),
-                    "error should mention channel_id mismatch: {error}",
-                );
-            },
-        );
-
-        fs::remove_file(&script_path).expect("script fixture should be removable");
-    }
-
-    #[test]
-    fn unit_run_live_s04_cli_task_lifecycle_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s04_cli_task_lifecycle_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s05_cli_escrow_settlement_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s05_cli_escrow_settlement_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_validate_live_s05_release_escrow_response_rejects_mismatched_escrow_id() {
-        let error = validate_live_s05_release_escrow_response(
-            "escrow-a",
-            "escrow-b",
-            "released",
-            "cli live s05 release-escrow",
-        )
-        .expect_err("mismatched escrow ids should fail");
-        assert!(
-            error.contains("mismatched escrow_id"),
-            "error should describe escrow-id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s06_cli_proof_verification_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s06_cli_proof_verification_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s07_cli_replay_protection_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s07_cli_replay_protection_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s08_cli_crash_recovery_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s08_cli_crash_recovery_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s08_cli_crash_recovery_probe_accepts_distinct_pre_post_message_ids() {
-        let script_path = unique_temp_script_path("kamn-e2e-cli-s08-success");
-        let script_source = r#"#!/usr/bin/env python3
-import os
-import sys
-
-command = sys.argv[1] if len(sys.argv) > 1 else ""
-agent_name = os.environ.get("KAMN_AGENT_NAME", "")
-
-if command == "send-message":
-    if agent_name.endswith("pre-send"):
-        sys.stdout.write("message_id=message-pre status=sent")
-    elif agent_name.endswith("post-send"):
-        sys.stdout.write("message_id=message-post status=sent")
-    else:
-        sys.stdout.write("message_id=message-fallback status=sent")
-elif command == "query-message":
-    message_id = sys.argv[-1] if len(sys.argv) > 0 else "message-fallback"
-    sys.stdout.write(f"message_id={message_id} status=sent")
-elif command == "health":
-    sys.stdout.write("status=ok")
-else:
-    sys.stderr.write("unsupported command")
-    sys.exit(2)
-"#;
-        write_executable_python_script(&script_path, script_source);
-
-        with_env_vars(
-            &[
-                (
-                    CLI_BINARY_ENV,
-                    Some(
-                        script_path
-                            .to_str()
-                            .expect("script path should be valid utf-8"),
-                    ),
-                ),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                run_live_s08_cli_crash_recovery_probe()
-                    .expect("distinct pre/post message IDs should pass continuity checks");
-            },
-        );
-
-        fs::remove_file(&script_path).expect("script fixture should be removable");
-    }
-
-    #[test]
-    fn unit_run_live_s09_cli_transport_failover_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-                (
-                    "KAMN_E2E_S09_FAILOVER_ENDPOINT",
-                    Some("http://localhost:8081"),
-                ),
-            ],
-            || {
-                let error = run_live_s09_cli_transport_failover_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s09_cli_transport_failover_probe_accepts_distinct_pre_post_message_ids() {
-        let script_path = unique_temp_script_path("kamn-e2e-cli-s09-success");
-        let script_source = r#"#!/usr/bin/env python3
-import os
-import sys
-
-command = sys.argv[1] if len(sys.argv) > 1 else ""
-agent_name = os.environ.get("KAMN_AGENT_NAME", "")
-
-if command == "send-message":
-    if agent_name.endswith("pre-send"):
-        sys.stdout.write("message_id=message-pre status=sent")
-    elif agent_name.endswith("post-send"):
-        sys.stdout.write("message_id=message-post status=sent")
-    else:
-        sys.stdout.write("message_id=message-fallback status=sent")
-elif command == "query-message":
-    message_id = sys.argv[-1] if len(sys.argv) > 0 else "message-fallback"
-    sys.stdout.write(f"message_id={message_id} status=sent")
-elif command == "health":
-    sys.stdout.write("status=ok")
-else:
-    sys.stderr.write("unsupported command")
-    sys.exit(2)
-"#;
-        write_executable_python_script(&script_path, script_source);
-
-        with_env_vars(
-            &[
-                (
-                    CLI_BINARY_ENV,
-                    Some(
-                        script_path
-                            .to_str()
-                            .expect("script path should be valid utf-8"),
-                    ),
-                ),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-                (
-                    "KAMN_E2E_S09_FAILOVER_ENDPOINT",
-                    Some("http://localhost:8081"),
-                ),
-            ],
-            || {
-                run_live_s09_cli_transport_failover_probe()
-                    .expect("distinct pre/post message IDs should pass continuity checks");
-            },
-        );
-
-        fs::remove_file(&script_path).expect("script fixture should be removable");
-    }
-
-    #[test]
-    fn unit_run_live_s10_cli_topology_coherence_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                (
-                    "KAMN_E2E_S10_PRIMARY_ENDPOINT",
-                    Some("http://localhost:8080"),
-                ),
-                (
-                    "KAMN_E2E_S10_SECONDARY_ENDPOINT",
-                    Some("http://localhost:8081"),
-                ),
-                (
-                    "KAMN_E2E_S10_TERTIARY_ENDPOINT",
-                    Some("http://localhost:8082"),
-                ),
-            ],
-            || {
-                let error = run_live_s10_cli_topology_coherence_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s10_cli_topology_coherence_probe_accepts_topology_query_continuity() {
-        let script_path = unique_temp_script_path("kamn-e2e-cli-s10-success");
-        let script_source = r#"#!/usr/bin/env python3
-import os
-import sys
-
-command = sys.argv[1] if len(sys.argv) > 1 else ""
-
-if command == "send-message":
-    sys.stdout.write("message_id=message-primary status=sent")
-elif command == "query-message":
-    message_id = sys.argv[-1] if len(sys.argv) > 0 else "message-primary"
-    sys.stdout.write(f"message_id={message_id} status=sent")
-elif command == "health":
-    sys.stdout.write("status=ok")
-else:
-    sys.stderr.write("unsupported command")
-    sys.exit(2)
-"#;
-        write_executable_python_script(&script_path, script_source);
-
-        with_env_vars(
-            &[
-                (
-                    CLI_BINARY_ENV,
-                    Some(
-                        script_path
-                            .to_str()
-                            .expect("script path should be valid utf-8"),
-                    ),
-                ),
-                (
-                    "KAMN_E2E_S10_PRIMARY_ENDPOINT",
-                    Some("http://localhost:8080"),
-                ),
-                (
-                    "KAMN_E2E_S10_SECONDARY_ENDPOINT",
-                    Some("http://localhost:8081"),
-                ),
-                (
-                    "KAMN_E2E_S10_TERTIARY_ENDPOINT",
-                    Some("http://localhost:8082"),
-                ),
-            ],
-            || {
-                run_live_s10_cli_topology_coherence_probe()
-                    .expect("topology query continuity should pass");
-            },
-        );
-
-        fs::remove_file(&script_path).expect("script fixture should be removable");
-    }
-
-    #[test]
-    fn unit_run_live_s11_cli_signer_rotation_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s11_cli_signer_rotation_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s12_cli_retention_deletion_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s12_cli_retention_deletion_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s13_cli_bridge_forwarding_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s13_cli_bridge_forwarding_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s14_cli_batch_merkle_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error =
-                    run_live_s14_cli_batch_merkle_probe().expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s15_cli_performance_smoke_probe_rejects_missing_binary() {
-        with_env_vars(
-            &[
-                (CLI_BINARY_ENV, Some("/definitely/missing/kamn-cli")),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                let error = run_live_s15_cli_performance_smoke_probe()
-                    .expect_err("missing binary should fail");
-                assert!(
-                    error.contains("failed to spawn"),
-                    "error should reflect spawn failure: {error}",
-                );
-            },
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s14_cli_batch_merkle_probe_accepts_distinct_batch_ids_and_final_proofs() {
-        let script_path = unique_temp_script_path("kamn-e2e-cli-s14-success");
-        let script_source = r#"#!/usr/bin/env python3
-import os
-import sys
-
-command = sys.argv[1] if len(sys.argv) > 1 else ""
-agent_name = os.environ.get("KAMN_AGENT_NAME", "")
-
-if command == "send-message":
-    if agent_name.endswith("-batch-a"):
-        sys.stdout.write("message_id=message-batch-a status=sent")
-    elif agent_name.endswith("-batch-b"):
-        sys.stdout.write("message_id=message-batch-b status=sent")
-    else:
-        sys.stdout.write("message_id=message-fallback status=sent")
-elif command == "query-message":
-    message_id = sys.argv[-1] if len(sys.argv) > 0 else "message-fallback"
-    sys.stdout.write(f"message_id={message_id} status=sent")
-elif command == "verify-proof":
-    message_id = sys.argv[-4] if len(sys.argv) >= 4 else "message-fallback"
-    block_height = sys.argv[-2] if len(sys.argv) >= 2 else "1"
-    sys.stdout.write(
-        f"message_id={message_id} verified=true finality=FINAL block_height={block_height}"
-    )
-else:
-    sys.stderr.write("unsupported command")
-    sys.exit(2)
-"#;
-        write_executable_python_script(&script_path, script_source);
-
-        with_env_vars(
-            &[
-                (
-                    CLI_BINARY_ENV,
-                    Some(
-                        script_path
-                            .to_str()
-                            .expect("script path should be valid utf-8"),
-                    ),
-                ),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-                ("KAMN_E2E_S14_AGENT_NAME", Some("kamn-e2e-cli-s14")),
-            ],
-            || {
-                run_live_s14_cli_batch_merkle_probe()
-                    .expect("distinct batch IDs with final proofs should pass");
-            },
-        );
-
-        fs::remove_file(&script_path).expect("script fixture should be removable");
-    }
-
-    #[test]
-    fn unit_validate_s14_cli_verify_proof_response_accepts_valid_payload() {
-        validate_s14_cli_verify_proof_response(
-            "message_id=message-1 verified=true finality=FINAL block_height=42",
-            "message-1",
-            "test helper",
-        )
-        .expect("valid S-14 proof payload should pass");
-    }
-
-    #[test]
-    fn unit_validate_s14_cli_verify_proof_response_rejects_mismatched_message_id() {
-        let error = validate_s14_cli_verify_proof_response(
-            "message_id=message-2 verified=true finality=FINAL block_height=42",
-            "message-1",
-            "test helper",
-        )
-        .expect_err("mismatched message_id should fail");
-        assert!(
-            error.contains("mismatched message_id"),
-            "error should mention message_id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s14_cli_verify_proof_response_rejects_unverified_payload() {
-        let error = validate_s14_cli_verify_proof_response(
-            "message_id=message-1 verified=false finality=FINAL block_height=42",
-            "message-1",
-            "test helper",
-        )
-        .expect_err("verified=false should fail");
-        assert!(
-            error.contains("verified=false"),
-            "error should mention verified contract: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s14_cli_verify_proof_response_rejects_non_final_finality() {
-        let error = validate_s14_cli_verify_proof_response(
-            "message_id=message-1 verified=true finality=PENDING block_height=42",
-            "message-1",
-            "test helper",
-        )
-        .expect_err("non-final finality should fail");
-        assert!(
-            error.contains("non-final finality"),
-            "error should mention finality contract: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s14_cli_verify_proof_response_rejects_zero_block_height() {
-        let error = validate_s14_cli_verify_proof_response(
-            "message_id=message-1 verified=true finality=FINAL block_height=0",
-            "message-1",
-            "test helper",
-        )
-        .expect_err("block_height=0 should fail");
-        assert!(
-            error.contains("block_height=0"),
-            "error should mention block-height contract: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s15_latency_budget_samples_accepts_within_budget_samples() {
-        validate_s15_latency_budget_samples(&[10, 20, 30], 80, 100, 25, 35, "test helper")
-            .expect("within-budget samples should pass");
-    }
-
-    #[test]
-    fn unit_validate_s15_latency_budget_samples_rejects_total_budget_violation() {
-        let error =
-            validate_s15_latency_budget_samples(&[10, 20, 30], 120, 100, 25, 35, "test helper")
-                .expect_err("total budget violation should fail");
-        assert!(
-            error.contains("total elapsed millis exceeded budget"),
-            "error should mention total budget: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s15_latency_budget_samples_rejects_p50_budget_violation() {
-        let error =
-            validate_s15_latency_budget_samples(&[10, 50, 90], 90, 200, 20, 100, "test helper")
-                .expect_err("p50 budget violation should fail");
-        assert!(
-            error.contains("p50 millis exceeded budget"),
-            "error should mention p50 budget: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s15_latency_budget_samples_rejects_p99_budget_violation() {
-        let error =
-            validate_s15_latency_budget_samples(&[10, 20, 90], 90, 200, 50, 80, "test helper")
-                .expect_err("p99 budget violation should fail");
-        assert!(
-            error.contains("p99 millis exceeded budget"),
-            "error should mention p99 budget: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_parse_s15_budget_env_u128_uses_default_when_env_missing() {
-        with_env_vars(&[("KAMN_E2E_S15_TOTAL_BUDGET_MS", None)], || {
-            let parsed = parse_s15_budget_env_u128(
-                "KAMN_E2E_S15_TOTAL_BUDGET_MS",
-                91,
-                "cli-scripted live s15 test helper",
-            )
-            .expect("missing env key should use default");
-            assert_eq!(parsed, 91);
-        });
-    }
-
-    #[test]
-    fn unit_parse_s15_budget_env_u128_parses_positive_env_value() {
-        with_env_vars(&[("KAMN_E2E_S15_TOTAL_BUDGET_MS", Some("143"))], || {
-            let parsed = parse_s15_budget_env_u128(
-                "KAMN_E2E_S15_TOTAL_BUDGET_MS",
-                91,
-                "cli-scripted live s15 test helper",
-            )
-            .expect("valid env key should parse");
-            assert_eq!(parsed, 143);
-        });
-    }
-
-    #[test]
-    fn unit_validate_s15_latency_budget_samples_accepts_exact_budget_boundaries() {
-        validate_s15_latency_budget_samples(&[10, 20, 30], 60, 60, 20, 30, "test helper")
-            .expect("equal total/p50/p99 budget boundaries should pass");
-    }
-
-    #[test]
-    fn unit_percentile_index_returns_expected_midpoint_index() {
-        assert_eq!(
-            crate::drivers::shared_helpers::percentile_index(3, 50),
-            1,
-            "len=3 and p50 should map to middle sample index",
-        );
-    }
-
-    #[test]
-    fn unit_percentile_index_clamps_percentile_above_hundred_to_last_index() {
-        assert_eq!(
-            crate::drivers::shared_helpers::percentile_index(3, 150),
-            2,
-            "percentiles above 100 should clamp to the last sample index",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s12_content_id_match_rejects_mismatch() {
-        let error = validate_s12_content_id_match("content-a", "content-b", "test step")
-            .expect_err("mismatched content ids should fail");
-        assert!(
-            error.contains("mismatched content_id"),
-            "error should mention content_id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s12_content_field_coherence_rejects_drift() {
-        let error = validate_s12_content_field_coherence(
-            "tombstoned",
-            "expired",
-            "lifecycle_state",
-            "test step",
-        )
-        .expect_err("field drift should fail");
-        assert!(
-            error.contains("lifecycle_state drift"),
-            "error should mention field drift: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s13_bridge_id_match_rejects_mismatch() {
-        let error = validate_s13_bridge_id_match("bridge-a", "bridge-b", "test step")
-            .expect_err("mismatched bridge ids should fail");
-        assert!(
-            error.contains("mismatched bridge_id"),
-            "error should mention bridge_id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s13_bridge_field_coherence_rejects_drift() {
-        let error =
-            validate_s13_bridge_field_coherence("forwarded", "stale", "bridge_status", "test step")
-                .expect_err("bridge field drift should fail");
-        assert!(
-            error.contains("bridge_status drift"),
-            "error should mention field drift: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_run_live_s11_cli_signer_rotation_probe_accepts_rotation_continuity() {
-        let script_path = unique_temp_script_path("kamn-e2e-cli-s11-success");
-        let script_source = r#"#!/usr/bin/env python3
-import os
-import sys
-
-command = sys.argv[1] if len(sys.argv) > 1 else ""
-agent_name = os.environ.get("KAMN_AGENT_NAME", "")
-payload = sys.argv[-1] if len(sys.argv) > 0 else ""
-primary_agent_name = os.environ.get("KAMN_E2E_S11_PRIMARY_AGENT_NAME", "kamn-e2e-cli-s11-primary")
-rotated_agent_name = os.environ.get("KAMN_E2E_S11_ROTATED_AGENT_NAME", f"{primary_agent_name}-rotated")
-stale_payload = os.environ.get("KAMN_E2E_S11_STALE_MESSAGE_PAYLOAD", "{\"message\":\"cli-scripted-live-s11-stale\"}")
-
-if command == "send-message":
-    if agent_name == primary_agent_name and payload == stale_payload:
-        sys.stderr.write("service_api_auth_replay_nonce_detected")
-        sys.exit(1)
-    if agent_name == primary_agent_name:
-        sys.stdout.write("message_id=message-primary status=sent")
-    elif agent_name == rotated_agent_name:
-        sys.stdout.write("message_id=message-rotated status=sent")
-    else:
-        sys.stdout.write("message_id=message-fallback status=sent")
-elif command == "query-message":
-    message_id = sys.argv[-1] if len(sys.argv) > 0 else "message-fallback"
-    sys.stdout.write(f"message_id={message_id} status=sent")
-else:
-    sys.stderr.write("unsupported command")
-    sys.exit(2)
-"#;
-        write_executable_python_script(&script_path, script_source);
-
-        with_env_vars(
-            &[
-                (
-                    CLI_BINARY_ENV,
-                    Some(
-                        script_path
-                            .to_str()
-                            .expect("script path should be valid utf-8"),
-                    ),
-                ),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-                (
-                    "KAMN_E2E_S11_PRIMARY_AGENT_NAME",
-                    Some("kamn-e2e-cli-s11-primary"),
-                ),
-                (
-                    "KAMN_E2E_S11_ROTATED_AGENT_NAME",
-                    Some("kamn-e2e-cli-s11-rotated"),
-                ),
-            ],
-            || {
-                run_live_s11_cli_signer_rotation_probe()
-                    .expect("signer-rotation continuity should pass");
-            },
-        );
-
-        fs::remove_file(&script_path).expect("script fixture should be removable");
-    }
-
-    #[test]
-    fn unit_run_live_s06_cli_proof_verification_probe_accepts_success_payload() {
-        let script_path = unique_temp_script_path("kamn-e2e-cli-s06-success");
-        let script_source = format!(
-            r#"#!/usr/bin/env python3
-import sys
-sys.stdout.write({payload:?})
-"#,
-            payload = "message_id=s06-live-proof block_height=1 finality=FINAL verified=true"
-        );
-        write_executable_python_script(&script_path, script_source.as_str());
-
-        with_env_vars(
-            &[
-                (
-                    CLI_BINARY_ENV,
-                    Some(
-                        script_path
-                            .to_str()
-                            .expect("script path should be valid utf-8"),
-                    ),
-                ),
-                ("KAMN_ENDPOINT", Some("http://localhost:8080")),
-            ],
-            || {
-                run_live_s06_cli_proof_verification_probe()
-                    .expect("success payload should pass verification probe");
-            },
-        );
-
-        fs::remove_file(&script_path).expect("script fixture should be removable");
-    }
-
-    #[test]
-    fn unit_run_cli_command_capture_stdout_returns_trimmed_stdout_on_success() {
-        let output = run_cli_command_capture_stdout(
-            "/bin/sh",
-            &["-c", "printf 'task_id=task-1 state=created'"],
-            "test helper",
-        )
-        .expect("successful command should return stdout");
-        assert_eq!(output, "task_id=task-1 state=created");
-    }
-
-    #[test]
-    fn unit_run_cli_command_capture_stdout_rejects_non_success_exit_status() {
-        let error = run_cli_command_capture_stdout("/bin/sh", &["-c", "exit 7"], "test helper")
-            .expect_err("non-success status should fail");
-        assert!(
-            error.contains("exit_status=7"),
-            "error should include failing exit status: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_run_cli_command_expect_failure_with_agent_name_returns_stderr() {
-        let output = super::run_cli_command_expect_failure_with_agent_name(
-            "/bin/sh",
-            &["-c", "echo replay >&2; exit 2"],
-            "test helper",
-            "probe",
-        )
-        .expect("stderr should be captured on expected failure");
-        assert_eq!(output, "replay");
-    }
-
-    #[test]
-    fn unit_run_cli_command_expect_failure_with_agent_name_rejects_success_status() {
-        let error = super::run_cli_command_expect_failure_with_agent_name(
-            "/bin/sh",
-            &["-c", "exit 0"],
-            "test helper",
-            "probe",
-        )
-        .expect_err("success status should be rejected");
-        assert!(
-            error.contains("unexpectedly succeeded"),
-            "error should mention unexpected success: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_parse_text_output_field_extracts_known_keys_and_missing_is_none() {
-        let output = "task_id=task-1 state=created";
-        assert_eq!(parse_text_output_field(output, "task_id"), Some("task-1"));
-        assert_eq!(parse_text_output_field(output, "state"), Some("created"));
-        assert_eq!(parse_text_output_field(output, "missing"), None);
-    }
-
-    #[test]
-    fn unit_validate_s07_replay_reason_marker_accepts_expected_marker() {
-        validate_s07_replay_reason_marker(
-            "operation failed: service_api_auth_replay_nonce_detected",
-            "test helper",
-        )
-        .expect("expected marker should be accepted");
-    }
-
-    #[test]
-    fn unit_validate_s07_replay_reason_marker_rejects_missing_marker() {
-        let error = validate_s07_replay_reason_marker("operation failed", "test helper")
-            .expect_err("missing marker should fail");
-        assert!(
-            error.contains("missing replay reason marker"),
-            "error should mention replay marker contract: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s08_message_receipt_fields_rejects_empty_message_id() {
-        let error = validate_s08_message_receipt_fields("message_id= status=sent", "test helper")
-            .expect_err("empty message_id should fail");
-        assert!(
-            error.contains("empty message_id"),
-            "error should mention message_id requirement: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_validate_s08_query_message_response_rejects_mismatched_message_id() {
-        let error = validate_s08_query_message_response(
-            "message_id=message-2 status=sent",
-            "message-1",
-            "test helper",
-        )
-        .expect_err("mismatched message_id should fail");
-        assert!(
-            error.contains("mismatched message_id"),
-            "error should mention message_id mismatch: {error}",
-        );
-    }
-
-    #[test]
-    fn unit_live_s07_probe_agent_suffix_is_non_empty_numeric() {
-        let suffix = super::live_s07_probe_agent_suffix();
-        assert!(!suffix.is_empty(), "suffix should be non-empty");
-        assert!(
-            suffix.chars().all(|character| character.is_ascii_digit()),
-            "suffix should be numeric: {suffix}",
-        );
-    }
-
-    #[test]
-    fn unit_cli_scripted_driver_debug_includes_live_toggle_field() {
-        let driver = CliScriptedDriver::with_runner(false, || Ok(()));
-        let debug = format!("{driver:?}");
-        assert!(debug.contains("CliScriptedDriver"));
-        assert!(debug.contains("live_execution_enabled"));
-    }
-
-    #[test]
-    fn spec_c00_live_disabled_driver_path_fails_closed_without_runner_invocation() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-        use std::sync::Arc;
-
-        let runner_calls = Arc::new(AtomicUsize::new(0));
-        let runner_calls_for_closure = Arc::clone(&runner_calls);
-        let driver = CliScriptedDriver::with_runner(false, move || {
-            runner_calls_for_closure.fetch_add(1, Ordering::SeqCst);
-            Ok(())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-01");
-        assert_eq!(
-            result.status, "fail",
-            "live-disabled S-01 must fail closed instead of reporting pass",
-        );
-        assert_eq!(
-            runner_calls.load(Ordering::SeqCst),
-            0,
-            "live runner should not execute when toggle is disabled",
-        );
-    }
-
-    #[test]
-    fn spec_c01_live_s04_driver_path_fails_closed_when_task_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s04 task probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-04");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-04 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c02_live_s06_driver_path_fails_closed_when_proof_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s06 proof probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-06");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-06 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c03_live_s02_driver_path_fails_closed_when_message_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s02 message probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-02");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-02 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c04_live_s03_driver_path_fails_closed_when_channel_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s03 channel probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-03");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-03 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c05_live_s05_driver_path_fails_closed_when_escrow_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s05 escrow probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-05");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-05 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c07_live_s07_driver_path_fails_closed_when_replay_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s07 replay probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-07");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-07 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c08_live_s08_driver_path_fails_closed_when_crash_recovery_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s08 crash-recovery probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-08");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-08 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c09_live_s09_driver_path_fails_closed_when_transport_failover_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s09 transport-failover probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-09");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-09 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c10_live_s10_driver_path_fails_closed_when_topology_coherence_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s10 topology-coherence probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-10");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-10 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c11_live_s11_driver_path_fails_closed_when_signer_rotation_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s11 signer-rotation probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-11");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-11 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c12_live_s12_driver_path_fails_closed_when_retention_deletion_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s12 retention-deletion probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-12");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-12 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c13_live_s13_driver_path_fails_closed_when_bridge_forwarding_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s13 bridge-forwarding probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-13");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-13 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c14_live_s14_driver_path_fails_closed_when_batch_merkle_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s14 batch-merkle probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-14");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-14 should fail closed on probe error",
-        );
-    }
-
-    #[test]
-    fn spec_c15_live_s15_driver_path_fails_closed_when_performance_smoke_probe_errors() {
-        let driver = CliScriptedDriver::with_runner(true, || {
-            Err("cli-scripted live s15 performance-smoke probe failed".to_owned())
-        });
-        let result = crate::drivers::HarnessDriver::execute(&driver, "S-15");
-        assert_eq!(
-            result.status, "fail",
-            "live-enabled S-15 should fail closed on probe error",
-        );
-    }
-}
+#[path = "cli_scripted_tests.rs"]
+mod cli_scripted_tests;
