+++ crates/kamn-e2e-harness/src/drivers/cli_scripted_tests/support.rs
@@ -1,0 +1,15 @@
+use super::*;
+
+#[path = "support/env_support.rs"]
+mod env_support;
+#[path = "support/script_fixture_support.rs"]
+mod script_fixture_support;
+#[path = "support/script_source_support.rs"]
+mod script_source_support;
+#[path = "support/update_support.rs"]
+mod update_support;
+
+pub(crate) use env_support::*;
+pub(crate) use script_fixture_support::*;
+pub(crate) use script_source_support::*;
+pub(crate) use update_support::*;
